+++ api/src/main/scala/com/tersesystems/blindsight/core/SourceInfoBehavior.scala
@@ -10,3 +10,19 @@ trait SourceInfoBehavior {
       enclosing: Enclosing
   ): Markers
 }
+
+object SourceInfoBehavior {
+
+  class Impl(
+      fileLabel: String,
+      lineLabel: String,
+      enclosingLabel: String
+  ) extends SourceCodeImplicits(fileLabel, lineLabel, enclosingLabel)
+      with SourceInfoBehavior {
+    override def apply(line: Line, file: File, enclosing: Enclosing): Markers = {
+      import com.tersesystems.blindsight.AST.BField
+      import com.tersesystems.blindsight.DSL._
+      Markers((line: BField) ~ file ~ enclosing)
+    }
+  }
+}
